+++ internal/server/controllers/tool.gen.go
@@ -10,6 +10,19 @@ import (
 	"github.com/langgenius/dify-plugin-daemon/pkg/entities/requests"
 )
 
+func InvokeTool(config *app.Config) gin.HandlerFunc {
+	type request = plugin_entities.InvokePluginRequest[requests.RequestInvokeTool]
+
+	return func(c *gin.Context) {
+		BindPluginDispatchRequest(
+			c,
+			func(itr request) {
+				service.InvokeTool(&itr, c, config.PluginMaxExecutionTimeout)
+			},
+		)
+	}
+}
+
 func ValidateToolCredentials(config *app.Config) gin.HandlerFunc {
 	type request = plugin_entities.InvokePluginRequest[requests.RequestValidateToolCredentials]
