+++ internal/server/controllers/tool.go
@@ -5,23 +5,7 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/langgenius/dify-plugin-daemon/internal/service"
-	"github.com/langgenius/dify-plugin-daemon/internal/types/app"
-	"github.com/langgenius/dify-plugin-daemon/pkg/entities/plugin_entities"
-	"github.com/langgenius/dify-plugin-daemon/pkg/entities/requests"
 )
-
-func InvokeTool(config *app.Config) gin.HandlerFunc {
-	type request = plugin_entities.InvokePluginRequest[requests.RequestInvokeTool]
-
-	return func(c *gin.Context) {
-		BindPluginDispatchRequest(
-			c,
-			func(itr request) {
-				service.InvokeTool(&itr, c, config.PluginMaxExecutionTimeout)
-			},
-		)
-	}
-}
 
 func ListTools(c *gin.Context) {
 	BindRequest(c, func(request struct {
