+++ internal/service/tool.gen.go
@@ -13,6 +13,23 @@ import (
 	"github.com/langgenius/dify-plugin-daemon/pkg/entities/tool_entities"
 )
 
+func InvokeTool(
+	r *plugin_entities.InvokePluginRequest[requests.RequestInvokeTool],
+	ctx *gin.Context,
+	max_timeout_seconds int,
+) {
+	baseSSEWithSession(
+		func(session *session_manager.Session) (*stream.Stream[tool_entities.ToolResponseChunk], error) {
+			return plugin_daemon.InvokeTool(session, &r.Data)
+		},
+		access_types.PLUGIN_ACCESS_TYPE_TOOL,
+		access_types.PLUGIN_ACCESS_ACTION_INVOKE_TOOL,
+		r,
+		ctx,
+		max_timeout_seconds,
+	)
+}
+
 func ValidateToolCredentials(
 	r *plugin_entities.InvokePluginRequest[requests.RequestValidateToolCredentials],
 	ctx *gin.Context,
